+++ database/migrations/2020_08_20_194929_create_marks_table.php
@@ -17,7 +17,7 @@ class CreateMarksTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('type_id');
             $table->unsignedBigInteger('studentNo');
-            $table->integer('markValue');
+            $table->float('markValue');
             $table->text('note')->nullable();
             $table->softDeletes();
             $table->timestamps();
